+++ packages/terra-action-header/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Revert components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 
 2.21.0 - (July 17, 2019)
 ------------------