+++ packages/terra-form-field/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Minor dependency version bump
 
 3.16.0 - (July 17, 2019)
 ------------------
@@ -17,7 +19,6 @@ Unreleased
 3.14.0  - (July 11, 2019)
 ------------------
 ### Changed
-* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 * Updated Docs around the requirement of terra-base and aggregate-translations.