+++ packages/terra-overlay/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Revert components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 
 3.21.0 - (July 17, 2019)
 ------------------